Fix reward shift bug and obs space dtype warning
- baseline=0.0 (not -1.0): sim failures now get shifted=0.0 and are skipped,
  not rewarded; partial/full successes get positive gradient as intended
- dtype= (not torch_dtype=) in from_pretrained to silence HF deprecation warning
- np.int64(step) in bracket_env obs to satisfy spaces.Discrete validation
- Added note in baseline cell: reward=0.00 is expected for untrained TinyLlama

+++ prodata/cad_gym/envs/bracket_env.py
@@ -75,7 +75,7 @@ class BracketDesignEnv(ProdataEnv):
             "load_kg": np.array([req.load_kg], dtype=np.float32),
             "extension_mm": np.array([req.extension_mm], dtype=np.float32),
             "max_cost_usd": np.array([req.max_cost_usd], dtype=np.float32),
-            "step": self._current_step,
+            "step": np.int64(self._current_step),
             "safety_factor": np.array([-1.0], dtype=np.float32),
             "cost_usd": np.array([-1.0], dtype=np.float32),
         }
